requirements.txt
pybit==5.3.0
Backtesting==0.3.3
ccxt==4.3.11
numpy==1.26.4
pandas==2.2.2
python-okx==0.2.8
plotly==5.21.0
requests==2.28.2
setuptools==69.2.0
websocket-client==1.8.0
